+++ client/src/ui/organisms/AddNewBudgetRecord.modal.jsx
@@ -1,4 +1,4 @@
-import React, { useContext } from 'react';
+import React, { useCallback, useContext } from 'react';
 import { Modal } from '../molecules/modal/Modal';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useForm, Controller } from 'react-hook-form';
@@ -12,10 +12,16 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
 import RootContext from '../../context/context';
 import { Button } from '../atoms/Button';
+import { useSnackbar } from 'notistack';
 
 export const AddNewBudgetRecordModal = ({ handleClose, ...props }) => {
   const context = useContext(RootContext);
   const { setOpenModal, setCategory, category } = context;
+  const { enqueueSnackbar } = useSnackbar();
+
+  const handleClick = useCallback((message, variant) => () => {
+    enqueueSnackbar(message, { variant: variant, anchorOrigin:{horizontal: "right", vertical: "bottom"}});
+  }, [enqueueSnackbar]);
 
   const schema = yup.object().shape({
     amount: yup
@@ -50,9 +56,13 @@ export const AddNewBudgetRecordModal = ({ handleClose, ...props }) => {
     (requestBody) => BudgetService.create({ requestBody }),
     {
       onSuccess: () => {
+        (handleClick("Budżet został zdefiniowany", "success"))()
         queryClient.invalidateQueries('budgetData');
         queryClient.invalidateQueries('categoriesData');
       },
+      onError:()=>{
+        (handleClick("Wystąpił nieoczekiwany błąd", "error"))()
+      }
     },
   );
   const onSubmit = (data) => {
@@ -76,30 +86,30 @@ export const AddNewBudgetRecordModal = ({ handleClose, ...props }) => {
     <Modal handleClose={handleCancel} {...props}>
       <form onSubmit={handleSubmit(onSubmit)}>
         <Controller
-          name='amount'
+          name="amount"
           control={control}
           rules={{ required: true, min: 0.01, max: 1000000 }}
           render={({ field }) => (
             <TextField
               {...field}
-              type='number'
+              type="number"
               inputProps={{ inputMode: 'numeric' }}
-              label='Kwota'
+              label="Kwota"
             />
           )}
         />
         <p>{errors.amount?.message}</p>
         <Controller
-          name='select'
+          name="select"
           control={control}
           render={({ field }) => (
             <FormControl fullWidth>
-              <InputLabel id='demo-simple-select-label'>Kategoria</InputLabel>
+              <InputLabel id="demo-simple-select-label">Kategoria</InputLabel>
               <Select
-                labelId='demo-simple-select-label'
-                name='select'
+                labelId="demo-simple-select-label"
+                name="select"
                 value={category}
-                label='Kategoria'
+                label="Kategoria"
                 onChange={handleChange}
                 {...field}
               >
@@ -118,17 +128,17 @@ export const AddNewBudgetRecordModal = ({ handleClose, ...props }) => {
         <p>{errors.select?.message}</p>
         <CardActions sx={{ display: 'flex', justifyContent: 'flex-end' }}>
           <Button
-            variant='outlined'
-            color='primary'
+            variant="outlined"
+            color="primary"
             sx={{ m: 2 }}
             onClick={(props) => handleCancel(props)}
           >
             Anuluj
           </Button>
           <Button
-            variant='contained'
-            color='primary'
-            type='submit'
+            variant="contained"
+            color="primary"
+            type="submit"
             disabled={!formState.isValid}
           >
             Zapisz
