+++ package.json
@@ -12,6 +12,7 @@
     "chart.js": "^3.7.1",
     "cypress": "^9.1.0",
     "luxon": "^2.3.0",
+    "notistack": "^2.0.3",
     "prettier": "2.5.1",
     "react-chartjs-2": "^4.0.1",
     "react-hook-form": "^7.27.1",
